classification-service/requirements.txt
fastapi
uvicorn
tensorflow
numpy
pillow
python-multipart
python-dotenv
python-jose[cryptography]